change shouldUseSingleTransactionView condition to work on transactions
